+++ tatk/policy/mle/loader.py
@@ -3,36 +3,21 @@ import pickle
 import torch
 import torch.utils.data as data
 from tatk.util.multiwoz.state import default_state
-from tatk.policy.loader.dataset import ActDataset
-from tatk.policy.vector.vector_multiwoz import MultiWozVector
+from tatk.policy.vector.dataset import ActDataset
 from tatk.util.dataloader.module_dataloader import ActPolicyDataloader
 
-class ActPolicyDataLoaderMultiWoz():
+class ActMLEPolicyDataLoader():
     
     def __init__(self):
-        root_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))))
-        voc_file = os.path.join(root_dir, 'data/multiwoz/sys_da_voc.txt')
-        voc_opp_file = os.path.join(root_dir, 'data/multiwoz/usr_da_voc.txt')
-        self.vector = MultiWozVector(voc_file, voc_opp_file)
+        self.vector = None
         
-        processed_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'processed_data')
-        if os.path.exists(processed_dir):
-            print('Load processed data file')
-            self._load_data(processed_dir)
-        else:
-            print('Start preprocessing the dataset')
-            self._build_data(root_dir, processed_dir)
-        
-    def _build_data(self, root_dir, processed_dir):
-        raw_data = {}
-        for part in ['train', 'val', 'test']:
-            raw_data[part] = ActPolicyDataloader(data_key=part, role='sys')
-        
+    def _build_data(self, root_dir, processed_dir):        
         self.data = {}
         for part in ['train', 'val', 'test']:
             self.data[part] = []
+            raw_data = ActPolicyDataloader(data_key=part, role='sys')
             
-            for turn in raw_data[part]:
+            for turn in raw_data:
                 state = default_state()
                 state['user_action'] = turn['context_dialog_act'][-1]
                 state['system_action'] = turn['context_dialog_act'][-2]
