+++ tatk/policy/mle/multiwoz/loader.py
@@ -1,0 +1,20 @@
+import os
+from tatk.policy.vector.vector_multiwoz import MultiWozVector
+from tatk.policy.mle.loader import ActMLEPolicyDataLoader
+
+class ActMLEPolicyDataLoaderMultiWoz(ActMLEPolicyDataLoader):
+    
+    def __init__(self):
+        root_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))))
+        voc_file = os.path.join(root_dir, 'data/multiwoz/sys_da_voc.txt')
+        voc_opp_file = os.path.join(root_dir, 'data/multiwoz/usr_da_voc.txt')
+        self.vector = MultiWozVector(voc_file, voc_opp_file)
+        
+        processed_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'processed_data')
+        if os.path.exists(processed_dir):
+            print('Load processed data file')
+            self._load_data(processed_dir)
+        else:
+            print('Start preprocessing the dataset')
+            self._build_data(root_dir, processed_dir)
+
